requirements.txt
ipwhois == 0.11.0
mechanize == 0.2.5
python_nmap == 0.4.3
paramiko == 1.15.2
pythonwhois == 2.4.3
requests == 2.7.0
shodan == 1.3.6
Pillow == 3.1.0
PyPDF2 == 1.25.1
beautifulsoup4 == 4.4.1
dnspython == 1.12.0
